Deck Builder: Deck Contents Display and Move To Deck Arrow Implemented.
Deck Contents are now displayed in the UI. Both DeckData and Deck Object classes have been implemented.

"Move to Deck arrow" button implemted. When clicked the selected card from the Storage will move to the deck if the deck side is < 20 cards and if the deck doesnt have 3 copies or that card. Fusion deck can only store 3 cards total.