SoundServer: Implementation of Sound Effects
SoundServer now offers a "PlayEffect" function to play a short sounds based on the current list of effects.

Applied effects to all the forms with sounds for Clicking buttons, Hover over elements; all these are based on when needed and what kind of sounds.

Also I took the time to refactor the LoadPages/LoadDeck in the DeckBuilder menu to be more efficients. Now reloading both when moving cards is faster.