BoardFrom Mock implemented and Card and Tile Object
First Mock draft for the Board Form. The constructor of this class will generate the Board Grid 13x18 tiles.

Tiles are their own object to referece to the picture boxes the represent each time.

Also create a Card object, this will hold the data of card on the board. A sample "Summon Card" function has been implemented to place a card into the board. ALL THIS IS STILL WIP.