Board: Card Info Display Implemented
On the Board when hovering a tile with a card in it a Card Info Panel will pop to display this card's info.

This panel will display either on top or bottom of the board based on where the tile is in relation to the form. This way the panel will never cover the tile that is being hover to.