BoardForm: Action Menu display implementation
Implementation of game states to flag which actions are available.

During the MainPhaseBoard player can hover cards on the board to see their info. Click on a card owned by the player opens the Action menu. The board is block until an action is selected (or selecting "Cancel") this is WIP and only the Cancel option was implemented.