Bug Fix: Missed clearing out the Move candidate list during the last tile movement.
During the Movement Sequence, when you use all the available crests to move. the player should not be able to click any tile for moving again. Because I forgot to clear the movement candidate list during this scenario, is was possible to still click the tiles candidates of the previous move.
Now the player can only either finish the move or cancel it.