BattleMenu Mock Implementation
Selecting Attack as an Action has been implemented. When doing so, the tiles with potential attack targers are highlighted. At this point the player can cancel the action or click a target to launch the Battle Menu.

All the UI pieces have been place but no functionalities have been implemented.