Battle: Defender Monster is Destroyed event implemented
When the player attack and the damage done reduces the LP of the defender monster, this monster is now destroyed and it is removed from the board and also flagged as "IsDiscarted" (I am using the term "discarted" to mark cards in the discard pile. Not all card go there by destruction.

Also, implemented a "End Battle" button. when the battle is over and the game continues, clicking this button will take the play back to the board  in the main phase game state.