Deck Symbol implementation + decks changes
Now a deck can contain a "Symbol", this will be the Dice Master on the board.
The DeckBuilder has been updated to allow selecting the desire symbol. 1 for each main attribute.

Also, I changed the dynamic of the deck builder. Player can only have up to 3 decks and these have been preset to Deck A B and C. Player cannot add or remove them.
Another main change is that the deckbuilder will only update the save file when exiting the menu. The "Exit" button has become "Save and Exit" instead. If there is not at LEAST 1 ready deck the save wont happen and the player cannot exit the form.