Roll Menu: Rolling Dice Mechanic Implemented
The Logic to randomly generate the Dice roll and determine if Dice/Cards can be Normal/Ritual Summon or Set has been implemented.

When a card is able to be summon/set buttons to perform the action are displayed.  (this buttons functions WIP)

If no cards can be summon/set display a "go to board" button (function WIP)

Dices that landed in non star/ritual crests are then added to the Player's Crest Pool and a display of the pool was added to the form.