Roll/Board: "Go to Board" and Set Card functionalities.
During the Roll Sequence, when a roll results in no cards to be Summon/Set a "Go to board" button appears. It has been implemented that when this button is clicked, the Roll menu closes and the Board opens in the MainPhaseBoardView State.

On the other hand, when a Card can be Set, the "Set Card" button takes the player to the board on the SetCard Phase. in this phase the (unoccupied) Summoning Tiles own by the player are marked in green and the player must select one of it to set the card to. Once one of this tiles is clicked, the card will be set on the board  and the MainPhaseBoardView phase starts.