Board: Dimension Highlight implemented (base)
When the player will summon a card, the player must "Dimension" the dice, meaning that it should select the tiles that this summon will turn into player own tiles. I implemented the base engine to hover a tile on the board and highlight the other tiles that will form the dimension - if the dimension is valid, color it Green, otherwise RED.

The next part of this is that when you click the tile itself, summon the monster on that tile.

After that, incorporate this to the "Summon" buttons in the Roll Menu Form to complete the summoning mechanic.

Solution was rebuild to include latest exe