Roll Menu: Sounds effects added and Refactor of Card Selector engine
Sounds effects added to all the elements with missing effects in the Roll Menu Form.

Also, the engine to select a card to move between the deck and the dice to roll pull has been refactor.
Before, the player has to click a card to "Select" it and then click the respective arrow button to move the card to the other side.
Now, hovering will "Select" the card and clicking it will perform the action to move the card to the other side. this  way the arrow buttons are irrelevant and removed.