Card "Type" string display fixed
Since now the Card "Type" is a enum instead of the string, the value of this removes all spaces and/or special characters (ex Divine-Beast now displays as "DivineBeast")

A TypeAsString accessor has been added that displays the correct string for Types with spaces of special characters. All forms that display card info have been updated.