Storage: Bundle cards implementation
The storage data has been modified to "bundle" the same cards together and show an amount.
For example, prior to this update, if the player had 10 copies of "Dark Magician" these 10 cards would take 10 storage spaces. NOW, they will take only 1 storage space and the Deck Builder will show an "Amount" label to track the number of copies.
Moving cards between decks and storage will update this label.

if a card's last copy was removed from storage, this card is completely remove for the storage slot until this card is added back again.