Get Dimension Tiles function moved to the Tile Object
I refactored this method to work from the Tile object itself. from here you can get the dimension tiles from Base Tile.

Also, the logic to validate is a Dimension List of Tiles is "Valid" (meaning it can be used for summoning) was refactored into a method that can be used for both players.