Dimension Class Implemented + Dimension Validation at Roll Menu added
Dimension Class is meant to encapsulate the information of a possible dimension option to summon a monster on the board.

This object will hold the array of tiles that will constitute the dimension, the dimension form and a valid flag that will be validation during object instantiation.

With this object implemented, the Roll Dice form can now preamble generate a list of "valid" dimensions. If there are no valid dimensions then the form will prevent the player from summoning monsters. This same logic will be used by the CPU player to prevent it to summon. ALSO, this will help with the AI for dimension selection. With the list of "valid" dimensions on hand, the CPU player can just "pick" any from the pool and perform the summon.