"End Turn" button visibility implemented
An "End Turn" button has been added to the UI. The logic of WHEN the button should be visible has been implemented, not the functionality of the button itself.

The button should only appear on the Main Phase while no other action is taking place.

Ex. button should be hidden while: Summoning/Set sequence is active, Player opened the action menu of a card on the board, Card View Mode is active. Once the player has completed the card action (move, attack) and the player returns to the Main Phase with no active action happening, the button will be visible again.