Roll Dice Menu: Return of "Unused" cards back to the deck when Summon/Set
As in today, the Roll Dice Menu would remove the 3 cards to be rolled from the deck. When a Summon/Set happens the form closes all the cards not summon/set are never sent back to the deck.
This updates handles that.
When a normal summon or set happens, the other 2 card not used are sent back to deck.
When a ritual summon happens, the Ritual Card has to be "Used" as well so only that 3rd card goes back to the deck.