End Turn Button + Roll Dice Menu for CPU started + OpponentAI Class
End Turn button has been implemented to clean up the One Turn data of the cards on the board and reduce the spellbound counters of monster. Also it triggers the CPU turn.

Once the CPU turn starts Roll Dice Menu for CPU is triggered. OpponentAI class has been created wit the implementation of the Roll Dice selection function. This function will return the cards that will be used for the Dice Roll.

As in right now this OpponentAI function gets the Dice Selection, the Dice Roll form performs the roll animation. add the crests and show the buttons that "allow" the seelection of summon/set/go to the board.