CPU Roll Dice Menu: "Go To Board" implementation
When the CPU rolls the dice and no summon/set is available the CPU will "Press" the "Go to the board" button.

At the board, the game will be at CPU turn state so the player cannot do anything.
No actions were implemented for the CPU other than end turn.

As extra, a "GetCopy" function has been added to the Deck Object. this will prevent that copying a deck uses an actual copy and not a copy of the object reference.