Implemented: Clicking owned card tile to open Action Menu + Clicking Cancel to close it
A player can now click a card they own in the board to open the action menu and clicking the "Cancel" option will close it.
The opponent player will see the Action Menu panel open to see the action of the TURN PLAYER but the opponent will not be able to trigger any action if they click any available option in the menu.