PvP Move Sequence Implemented
TURN PLAYER can now use the Move option in the Action Menu. When this happens the Move Menu displays with the "Finish" and "Cancel" option plus a label to display the number of available [MOV] Crests.

When clicking "Cancel" the card will reset to its original position and the Move Menu will close returning the Main Phase.
Option "Finish" becomes enable once the Player performs the first move sequence. When clicking "Finish" the Card in question will be placed in the destination tile and the [MOV] crest count will be updated to reflect the number of crests used.

The label that displays the available crests will turn RED COLOR if the amount < Move Cost of the Card. Thus the Player wont be able to perform any other move.

The opponent's UI is being properly displayed while TURN PLAYER  is taking action, and the opponent will NOT be able to interact with any of the UI while this is happening.