Attack Targetting implemented
When the TURN PLAYER selects the Attack Action, the UI will display the Attack Menu with the only option to "Cancel". Clicking "Cancel" will reset the UI back to the Main Phase.
When the Attack Menu opens, the valid attack target candidates will have GREEN color highlight. These are opponent cards only.

When the TURN PLAYER clicks a valid attack target candidate the Battle Menu appears and the game enters the Battle Phase.