Monster vs Monster PvP implementation complete
TURN PLAYER (The attacker) has the option to Attack (duh) and if he/she has Attribute Advantage the adv submenu will appear to allow it to add [ATK] on get the ATK bonus.

The defender player will have the option to "Defend" or "Pass". The defend option will not be available if the player does not have enough [DEF] to pay the defend cost of the monster.
The defender player will also have access the the adv bonus to increate the DEF bonus for its monster.

While the opposite opponent hasn't make a final Attack/Defense choice, a message "Waiting for opponent" will be visible. If the opponent made its move while the User player hasn't, the message will update to "Opponent is ready to battle!"

Once both player have made their move, the battle calculation will take place. The final damage will be calculated as follow: Damage = (Attackers Base ATK + (BonusCrestsUsed * 200)) - (Defenders Base DEF + (BonusCrestsUsed * 200);
If the final damage is enough to reduce the LP of the defender monster to 0, the monster will be destroyed and and left damage will be inflicted to the defender's player's Symbol.
Reducing the Symbol's LP to 0 will trigger the TMP game over screen.

For the Advantage UI, each player can use up to 5 bonus crests for a max total of an extra 1000 ATK/DEF points (as long as the player has enough crests to pay)

If the final damage of a battle results in 0, nothing happens, battle is over.

When the battle end and the game continues, the TURN PLAYER has the control to click the End Battle button. Then game resumes back to the Main Phase