Start Screen redesigned
Form still prompts user to a "Start Game" button. When this is pressed, the form will check if a save file exists. If it does, form will read the player base info to preview on the Load Game submenu and set the option selection to 'Load Game File".
User can still switch selection to new game, but starting a new game will override the save file.

If a save file does NOT exist the option "Load Game File" will simply not be visible and thus the user must proceed to start a new game.