Effect Object created and implemented the first sample effect
I created the base draft of the Effect Object. The PvPBoard will have a list of "Active Effects". When a card activates an effect this object will be created and added to the list. When an effect activates the proper function will be executed based on the "Effect ID" of the effect.

I also did a UI change on the Stats Label for the Cards on board. Now it uses 2 labels, 1 for ATK and 1 for DEF. now that these are separated I can change their color based on the stat bonus.

I also refactored the Tile functions to update the UI on changes. Now 1 master "UpdateTileUI" will handle this.