Simple OnSummon Effect Implemented
I used "M-Warrior #1" s effect as the base sample for a OnSummon effect. When this card is summon a Panel will prompt that the Card summoned has a OnSummon effect then after 2 secs it dessapears and the effect takes place. This effect does not target so it just executes as is.