Removal of Continuous Effects upon Origin Card destruction implemented
When a monster is destroyed by battle that had a Continuous Effect active on the board, the effect will be "reverted" to all the cards affected by it and then it will be removed from the "Active Effects" list so it is disregarded from now on.

A new "DestroyCard()" method was create to encapsulate all the instance of card destruction later in dev. This method will handle verifying if a card had an active effect upon destruction and thus trigger the removal.

To test this, we tested the destruction of "Thunder Dragon" while another "Thunder Dragon" was on the field.