Implemented a "EffectsAreImplemented" flag
Due to the way the logic validates if a card has an effect to activate on summon, any card with OnSummon or Continuous Effect in the DB will attempt to activate its effect. This will cause exception for cards in which I have not actually implement their effects in code. this flag can be set to filter all those cards whose effects are not implemented.