Tile Hovering of any kind will no longer be send to the server as an action.
Hovering is used to set the _CurrentlySelectedTile, to allow the NON TURN PLAYER to set this object reference, the Clicking Tile Event will perform the hovering for that player and thus continue with the intented action.

Since the NON TURN PLAYER cannot see in action how the TURN PLAYER is moving the mouse, the label banner that was used to describe the summon action will now be a generic "action instructions label" that will describe what the TURN PLAYER is doing for the opposite player.