Game Over sequence disconnects from server automatically
When a player wins the match and the game over panel shows up, the Board Form will send a message to the server to notify the game over. this message will be forwards back to the client to allow the client listening thread to stop and disconnect from the server. Now at this point the player will still be in the PvPBoard but not longer connected to the server. From here all the player has to do is click the "Exit" button to return to the PvPMenu ready to join another match