Bug Fix: Fix the deck modification error during a pvp match
When the PvPMenu creates a match, the match uses the actual Deck object that stores the deck info in the game. The match needs to use a dummy copy of the object so it can be modified without actually modifying the real deck. The fix is to use the Deck.GetCopy() to use that dummy copy