"Field Type" implementation into the Tile Object
Field Types can be set into a Tile. The board was updated to have a UI to display the Field Type of the Tile being hover over.
The tile with a field type set will display the custom image field texture. Moving sequence has been updated to properly update the tile selection and tile reset to keep consistency of the Tile Image/Backcolor.

A tile with a field type set that is occupied by a card will show the card image, if the card moves out of the tile, the field image will display agian.