Set Card sequence completed for PvP mode.
Setting cards were not working with the current code migrated from the original BoardForm. Now the RollDiceMenu will properly show up the "Set" button when a dice roll allows the card to be set. It will also restrict the set if the TURN PLAYER does not have unoccupied Summon Tiles. When this happens, a warning label will show on the TURN PLAYER's side to warn the player.

Clicking "Set" will take the player to the board and allow the TURN PLAYER to select a candidate tile to set the card into. The candidate tiles will be displayed as Green tiles.

Once the TURN PLAYER clicks a candidate tile, the card will be set into the board facedown. From here the board goes into the Main Phase.