"No Input" Game state added
During a Summon/Set there is enough time for the turn player to click on the board again before the next game state change. if the user click on the board the game will try another summon/set.

To fix this, I created a "NOINPUT" game state that lock the interaction the moment the original click is made.