Initial Polymerization activation sequence implemented
Effect Requirements validation for Polymerization has been implemented.
The card can only be activated if:
-Owner's Fusion Deck contains cards AND
-The Fusion Materials of at least ONE of those cards exists on the board under the Owner's control.

WHen the activation requirements if being validated, AND there are Fusion Cards "Ready" to be fusion summon, those ready cards are added to a list of _FusionCardsReadyForFusion to be used in the Fusion Card Selection Menu.
Because the opponent player DOES NOT run the effect activation requirements validation we have to send the list of _FusionCardsReadyForFusion so the opponent's UI can follow along.

Once the TURNPLAYER clicks "Activate" the Fusion Card Selection Menu will show up and display the candidates ready for fusion summon.

Next step is to make those candidates "hidden" for the opponent player and start implementing clicking on any of the "Fusion Summon" buttons to complete the fusion summon sequence.