Attribute Advantage Panel fix
When a player has a Attribute advantage, their battle menu UI will show the Bonus Crests panel. This panel was missing setting up the bonus crests count to the default 0. So if you happen to use bonus crests then the next battle the UI will start with that same amount rather then 0. This has been fixed.