Attack Range Implemented and UI for attack targeting updated
The Card Object now supports "Attack Range" value. Starts at 1 by default.

During the Action Menu when the player clicks "Attack" the tiles within the attack range will be calculated and displayed as highlighted.
Tiles with cards that can be target for an attack within the attack range will be mark with the Attack Target Icon.
The rest of the battle phase will continue as normal.
When canceling the attack target selection or when entering the Battle Phase the UI will reset all tiles that were highlighted during this process.