Shinato's Continuous effect implemented
This Continuous effect does do anything on activation. It only reacts to Monster Destruction by battle, therefore this change includes the reaction mechanic to this particular event.

In order to get this done I needed to do a "small" refactor on the BattleCalculation() to incorporate the check of the Monster Destruction Event. Resolution for these effects SHOULD NOT move game states.

Shinato's reaction effect will inflict damage to the opponent's Symbol equal to the destroyed monster's original attack. Since this reaction can reduce the Symbol's LP to 0, the BattleCalculation() will do a LP check on the symbols before returning to the MainPhase. If a Symbol's LP was reduced to 0 then it will trigger the Game Over.

A new additon to the UI is the implementation of "ReactionNotification" panel. When certain card effects have either continuous events that will trigger actions not easily caught in the UI, a Notification Panel will show on the top right corner of the Board to show what reaction is happening.
I added this Notification Panel to Shinato's effect to emphasize that the second set of damage inflicted to the symbol was due to shinato's effect and not just regular battle damage.
I also added this to BLS when calculating battle damage, before the final damage is shown, the notification panel will show alerting the players that BLS' effect modified the damage calculated.