Normal Summon "Animation" added
When a monster is normal summoned, a card preview is displayed, banishes, then the card is placed in the summon tile and it is hover to drag the players eyes to it. Once done the tile reloads to clear the hover and this happens during the duration of the normal summon SFX and the timing is perfect.