"Default Deck" flag implemented
Now we can mark a deck in the Deck Manager as "Default" (only 1 deck can be marked at a time)
This field is now stored in the save file and can be read/write.
The PvPMenu will auto select the deck with the default flag to the user doesnt have to selected each time it opens the PvPMenu.