Prize Card Implementation
The winner of the match will obtain 1 card at random.
I might update this to pull a based based on a probability table but that will have to be a seperate ticket if I do it.
As in right now the prize card will be ANY at random from the card database.