Profile Menu Form initial Implementation
This form can be open from the Player Planel in the Main Menu.
This will display more datailed info of the player.
This commit contains 2 settings: Change Player Name / Change Avatar
Functionality of these are implemented but the SaveFileManager does not save the Avatar ID/Experice/level data in the save file yet.

that will be next step