Equip Overlay Icon display while hovering implemented
Now when hovering tiles in the board (during the Board View Mode/Main Phase) if a player hovers over a face up Equip Spell, the card this equip is equipped to will display the Equip overlay icon. This will help to identify which card is the equip spell linked to.