Trigger Effect Selector UI implementation
Now during a summon of a monster, when a player has a Trap Card with a (TRIGGER) effect that can be activated; a prompt will be displayed for both players. The TURNPLAYER will get a message that the opponent player "is thinking" and no actions can be taken. Meanwhile the opponent player has a UI to select any of the trap cards that can be activated (with cost met) and this player has the choice to either cancel and not activate anything or select a card for activation.